refactor(backend): split ai summary and quiz runners
